// D4St-Spa/src/Models/DTOs/ISkillBarDTO.ts
import { ISkillWithImageDTO } from './ISkillWithImageDTO';

export interface ISkillBarDTO {
  skill1: ISkillWithImageDTO;
  skill2: ISkillWithImageDTO;
  skill3: ISkillWithImageDTO;
  skill4: ISkillWithImageDTO;
  skill5: ISkillWithImageDTO;
  skill6: ISkillWithImageDTO;
  skill7: ISkillWithImageDTO;
  skill8: ISkillWithImageDTO;
}